requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
SQLAlchemy==2.0.23
alembic==1.12.1

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database drivers
aiosqlite==0.19.0

# Environment variables
python-dotenv==1.0.0

# UUID support
uuid==1.30

# JSON handling
orjson==3.9.10

# Google Gemini AI (keeping existing integration)
google-generativeai==0.8.5
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
